nlp/requirements.txt
numpy==1.21.2
spacy==3.1.3
websockets==10.0
google-api-python-client==2.22.0
google_auth_oauthlib==0.4.6